requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
Pillow>=10.0.0
ultralytics>=8.1.0
pytesseract>=0.3.10
moviepy>=1.0.3
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.1.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
